refactor: Move NodeId to model folder
