+++ src/index.js
@@ -1,6 +1,6 @@
 export { default as Atviseproject } from './lib/config/Atviseproject';
 export { default as ProjectConfig } from './config/ProjectConfig';
-export { default as NodeId } from './lib/server/NodeId';
+export { default as NodeId } from './lib/model/opcua/NodeId';
 export { default as Transformer, TransformDirection } from './lib/transform/Transformer';
 export { default as PartialTransformer } from './lib/transform/PartialTransformer';
 export { default as SplittingTransformer } from './lib/transform/SplittingTransformer';
