+++ src/lib/server/Types.js
@@ -1,5 +1,5 @@
 import { DataType } from 'node-opcua';
-import NodeId from '../server/NodeId';
+import NodeId from '../model/opcua/NodeId';
 
 /**
  * Special, atvise releated OPC-UA type.
