+++ test/src/lib/server/ReadStream.spec.js
@@ -1,6 +1,6 @@
 import { StatusCodes } from 'node-opcua';
 import expect from '../../../expect';
-import NodeId from '../../../../src/lib/server/NodeId';
+import NodeId from '../../../../src/lib/model/opcua/NodeId';
 import ReadStream from '../../../../src/lib/server/ReadStream';
 
 /** @test {ReadStream} */
